feat(protocols): architecture-audit/1 extract package
