+++ packages/indexed-connectors/pyproject.toml
@@ -8,6 +8,7 @@ authors = [
 ]
 requires-python = ">=3.11"
 dependencies = [
+    "indexed-protocols",
     "indexed-utils",
     "indexed-parsing",
     "orjson>=3.10.0",
@@ -22,6 +23,7 @@ dependencies = [
 ]
 
 [tool.uv.sources]
+indexed-protocols = { workspace = true }
 indexed-utils = { workspace = true }
 indexed-parsing = { workspace = true }
