+++ packages/indexed-core/pyproject.toml
@@ -8,6 +8,7 @@ authors = [
 ]
 requires-python = ">=3.11"
 dependencies = [
+    "indexed-protocols",
     "indexed-utils",
     "indexed-connectors",
     "faiss-cpu>=1.11.0",
@@ -18,6 +19,7 @@ dependencies = [
 ]
 
 [tool.uv.sources]
+indexed-protocols = { workspace = true }
 indexed-utils = { workspace = true }
 indexed-connectors = { workspace = true }
