+++ packages/indexed-protocols/pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "indexed-protocols"
+version = "0.1.0"
+description = "Shared connector protocols and DTOs for indexed"
+requires-python = ">=3.11"
+dependencies = ["pydantic>=2.0.0"]
+
+[build-system]
+requires = ["hatchling", "hatch-una"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/protocols"]
