+++ packages/indexed-protocols/src/protocols/__init__.py
@@ -1,0 +1,19 @@
+from protocols.connectors import BaseConnector, DocumentConverter, DocumentReader
+from protocols.metadata import ConnectorMetadata
+from protocols.models import (
+    PhasedProgressCallback,
+    ProgressCallback,
+    ProgressUpdate,
+    SourceConfig,
+)
+
+__all__ = [
+    "BaseConnector",
+    "ConnectorMetadata",
+    "DocumentConverter",
+    "DocumentReader",
+    "PhasedProgressCallback",
+    "ProgressCallback",
+    "ProgressUpdate",
+    "SourceConfig",
+]
